Initially got things working
