+++ napari/_vispy/overlays/interaction_box.py
@@ -1,14 +1,56 @@
 import numpy as np
+from skimage.transform import AffineTransform, SimilarityTransform
 from vispy.scene.visuals import Compound, Line, Markers
+
+from ._interaction_box_constants import Box
+
+
+def inside_boxes(boxes):
+    """Checks which boxes contain the origin. Boxes need not be axis aligned
+
+    Parameters
+    ----------
+    boxes : (N, 8, 2) array
+        Array of N boxes that should be checked
+
+    Returns
+    -------
+    inside : (N,) array of bool
+        True if corresponding box contains the origin.
+    """
+
+    AB = boxes[:, 0] - boxes[:, 6]
+    AM = boxes[:, 0]
+    BC = boxes[:, 6] - boxes[:, 4]
+    BM = boxes[:, 6]
+
+    ABAM = np.multiply(AB, AM).sum(1)
+    ABAB = np.multiply(AB, AB).sum(1)
+    BCBM = np.multiply(BC, BM).sum(1)
+    BCBC = np.multiply(BC, BC).sum(1)
+
+    c1 = 0 <= ABAM
+    c2 = ABAM <= ABAB
+    c3 = 0 <= BCBM
+    c4 = BCBM <= BCBC
+
+    inside = np.all(np.array([c1, c2, c3, c4]), axis=0)
+
+    return inside
 
 
 class VispyInteractionBox:
-    def __init__(self, viewer):
+    def __init__(self, viewer, parent=None, order=0):
 
         self._viewer = viewer
-        self.node = Compound([Line(), Markers()])
-
+        self._interaction_box = viewer.interaction_box
+        self.node = Compound([Line(), Markers()], parent=parent)
+        self.node.order = order
         self._on_interaction_box_change()
+        self.initialize_mouse_events(viewer)
+        self._interaction_box.events.points.connect(
+            self._on_interaction_box_change
+        )
 
     @property
     def marker_node(self):
@@ -31,7 +73,7 @@ class VispyInteractionBox:
             edge_color,
             pos,
             width,
-        ) = self._viewer.overlay.interaction_box._compute_vertices_and_box()
+        ) = self._viewer.interaction_box._compute_vertices_and_box()
 
         if vertices is None or len(vertices) == 0:
             vertices = np.zeros((1, self._viewer.dims.ndisplay))
@@ -55,5 +97,208 @@ class VispyInteractionBox:
             width = 0
         else:
             pos = pos + 0.5
-
         self.line_node.set_data(pos=pos, color=edge_color, width=width)
+
+    def initialize_mouse_events(self, layer):
+        """Adds event handling functions to the layer"""
+
+        @layer.mouse_move_callbacks.append
+        def mouse_move(layer, event):
+            if (
+                not self._interaction_box.show
+                or self._interaction_box._box is None
+            ):
+                return
+
+            box = self._interaction_box._box
+            coord = event.position
+            distances = abs(box - coord)
+
+            # Get the vertex sizes
+            sizes = self._interaction_box._vertex_size / 2
+
+            # Check if any matching vertices
+            matches = np.all(distances <= sizes, axis=1).nonzero()
+            if len(matches[0]) > 0:
+                self._interaction_box._selected_vertex = matches[0][-1]
+                # Exclde center vertex
+                if self._interaction_box._selected_vertex == Box.CENTER:
+                    self._interaction_box._selected_vertex = None
+            else:
+                self._interaction_box._selected_vertex = None
+            # self.events.points_changed()
+
+        @layer.mouse_drag_callbacks.append
+        def mouse_drag(layer, event):
+            if not self._interaction_box.show:
+                return
+
+            # Handling drag start, decide what action to take
+            self._set_drag_start_values(layer, event.position)
+            drag_callback = None
+            final_callback = None
+            if self._interaction_box._selected_vertex is not None:
+                if self._interaction_box._selected_vertex == Box.HANDLE:
+                    drag_callback = self._on_drag_rotation
+                    final_callback = self._on_final_tranform
+                    yield
+                else:
+                    self._interaction_box._fixed_vertex = (
+                        self._interaction_box._selected_vertex + 4
+                    ) % Box.LEN
+                    drag_callback = self._on_drag_scale
+                    final_callback = self._on_final_tranform
+                    yield
+            else:
+                if (
+                    self._interaction_box._box is not None
+                    and self.show
+                    and inside_boxes(
+                        np.array([self._box - self._drag_start_coordinates])
+                    )[0]
+                ):
+                    drag_callback = self._on_drag_translate
+                    final_callback = self._on_final_tranform
+
+                    yield
+                else:
+                    self._interaction_box.points = None
+                    drag_callback = self._on_drag_newbox
+                    final_callback = self._on_end_newbox
+                    yield
+            # Handle events during dragging
+            while event.type == 'mouse_move':
+                if drag_callback is not None:
+                    drag_callback(layer, event)
+                yield
+
+            if final_callback is not None:
+                final_callback(layer, event)
+
+            self._clear_drag_start_values()
+
+    def _set_drag_start_values(self, layer, position):
+        """Gets called whenever a drag is started to remember starting values"""
+
+        self._drag_start_coordinates = np.array(position)
+        self._drag_start_box = np.copy(self._interaction_box._box)
+        if self._interaction_box._box is not None:
+            self._drag_start_angle = self._interaction_box.angle
+        self._drag_angle = 0
+        self._drag_scale = [1.0, 1.0]
+
+    def _clear_drag_start_values(self):
+        """Gets called at the end of a drag to reset remembered values"""
+
+        self._drag_start_coordinates = None
+        self._drag_start_box = None
+        self._drag_start_angle = None
+        self._drag_angle = 0
+        self._drag_scale = [1.0, 1.0]
+
+    def _on_drag_rotation(self, layer, event):
+        """Gets called upon mouse_move in the case of a rotation"""
+        center = self._drag_start_box[Box.CENTER]
+        new_offset = np.array(layer.world_to_data(event.position)) - center
+        new_angle = -np.degrees(np.arctan2(new_offset[0], -new_offset[1])) - 90
+
+        if np.linalg.norm(new_offset) < 1:
+            self._drag_angle = 0
+        elif self._fixed_aspect:
+            self._drag_angle = (
+                np.round(new_angle / 45) * 45 - self._drag_start_angle
+            )
+        else:
+            self._drag_angle = new_angle - self._drag_start_angle
+
+        tform1 = SimilarityTransform(translation=-center)
+        tform2 = SimilarityTransform(rotation=-np.radians(self._drag_angle))
+        tform3 = SimilarityTransform(translation=center)
+        transform = tform1 + tform2 + tform3
+        self._box = transform(self._drag_start_box)
+        self.events.points_changed()
+        self.events.transform_changed_drag(transform=transform)
+
+    def _on_drag_scale(self, layer, event):
+        """Gets called upon mouse_move in the case of a scaling operation"""
+
+        # Transform everything in axis-aligned space with fixed point at origin
+        center = self._drag_start_box[self._fixed_vertex]
+        transform = SimilarityTransform(translation=-center)
+        transform += SimilarityTransform(
+            rotation=np.radians(self._drag_start_angle)
+        )
+        coord = transform(np.array(layer.world_to_data(event.position)))[0]
+        drag_start = transform(self._drag_start_box[self._selected_vertex])[0]
+        # If sidepoint of fixed aspect ratio project offset onto vector along which to scale
+        # Since the fixed verted is now at the origin this vector is drag_start
+        if self._fixed_aspect or self._selected_vertex % 2 == 1:
+            offset = coord - drag_start
+            offset_proj = (
+                np.dot(drag_start, offset) / (np.linalg.norm(drag_start) ** 2)
+            ) * drag_start
+
+            # Prevent numeric instabilities
+            offset_proj[np.abs(offset_proj) < 1e-5] = 0
+            drag_start[drag_start == 0.0] = 1e-5
+
+            scale = np.array([1.0, 1.0]) + (offset_proj) / drag_start
+        else:
+            scale = coord / drag_start
+
+        # Apply scaling
+        transform += AffineTransform(scale=scale)
+
+        # Rotate and translate back
+        transform += SimilarityTransform(
+            rotation=-np.radians(self._drag_start_angle)
+        )
+        transform += SimilarityTransform(translation=center)
+        self._box = transform(self._drag_start_box)
+        self.events.points_changed()
+        self.events.transform_changed_drag(transform=transform)
+
+    def _on_drag_translate(self, layer, event):
+        """Gets called upon mouse_move in the case of a translation operation"""
+
+        offset = (
+            np.array(layer.world_to_data(event.position))
+            - self._drag_start_coordinates
+        )
+
+        transform = SimilarityTransform(translation=offset)
+        self._box = transform(self._drag_start_box)
+        self.events.points_changed()
+        self.events.transform_changed_drag(transform=transform)
+
+    def _on_final_tranform(self, layer, event):
+        """Gets called upon mouse_move in the case of a translation operation"""
+
+        self.events.transform_changed_final()
+
+    def _on_drag_newbox(self, layer, event):
+        """Gets called upon mouse_move in the case of a drawing a new box"""
+
+        self._interaction_box.points = np.array(
+            [
+                self._drag_start_coordinates,
+                np.array(event.position),
+            ]
+        )
+        self.show = True
+        self.show_handle = False
+        self.show_vertices = False
+        self._interaction_box.selection_box_drag = self._interaction_box._box[
+            Box.WITHOUT_HANDLE
+        ]
+
+    def _on_end_newbox(self, layer, event):
+        """Gets called upon mouse_move in the case of a drawing a new box"""
+        self.show = False
+        self.show_handle = True
+        self.show_vertices = True
+        # self.points = None
+        if self._interaction_box._box is not None:
+            self._interaction_box.selection_box_final = (
+                self._interaction_box._box[Box.WITHOUT_HANDLE]
+            )
